Criando busca por id do Tecnico

+++ udemy/angular_spring_boot/helpdesk/src/main/java/com/aug/helpdesk/repositories/TecnicoRepository.java
@@ -1,11 +1,9 @@
 package com.aug.helpdesk.repositories;
 
-import java.math.BigInteger;
-
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.aug.helpdesk.domain.Tecnico;
 
-public interface TecnicoRepository extends JpaRepository<Tecnico, BigInteger> {
+public interface TecnicoRepository extends JpaRepository<Tecnico, Integer> {
 
 }
